fix(Cache acc info): Changes to EntryViewModel
fix(Cache acc info):
Using the IO Dispatcher for the IO related operation and flattening the nested coroutines which ran under the same scope fixed the issue.

feat(Cache acc info):
Caches the Profile name and the email ID of the user. Profile picture appears as the navigation icon, although the size is'nt ideal for now.