fix: Sign In - Orientation Change
fix: Sign In -  Orientation Change:
Current implementation does not care  if the caching process is completed successfully or not. When there is an orientation  change and when the user has signed in and is waiting for the MainActivity to load,  the GoogleSignIn Instance  of the recreated activity will state that the sign in has been already done and the EntryActivity will  skip the incomplete caching process.